Refactor `nos/server` and `nos/client` modules for better extensibility.
 - fixed gRPC protocol errors for python3.8 and python3.7 interop
 - `nos/server`:
    - `docker`: Implements the `DockerRuntime` runtime daemon
    - `runtime`: Implements the `InferenceServiceRuntime`
    - `service`: Implements the gRPC  `InferenceService` server
 - `nos/client` is now a package with submodules `grpc` and `http`.
