+++ nos/client/__init__.py
@@ -1,0 +1,3 @@
+from nos.client.exceptions import NosClientException
+from nos.client.grpc.client import InferenceClient
+from nos.constants import DEFAULT_GRPC_PORT
