--- nos/client/grpc/__init__.py
@@ -1,2 +1,0 @@
-from nos.client.exceptions import NosClientException
-from nos.client.grpc.client import DEFAULT_GRPC_PORT, InferenceClient
