--- nos/experimental/grpc/__init__.py
@@ -1,2 +1,0 @@
-from nos.experimental.grpc.protoc import import_module  # noqa F401
-from nos.constants import DEFAULT_GRPC_PORT  # noqa F401
