+++ nos/server/__init__.py
@@ -1,2 +1,3 @@
 from .service import InferenceService  # noqa F401
 from .runtime import InferenceServiceRuntime  # noqa F401
+from .docker import DockerRuntime  # noqa F401
